requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.19
transformers>=4.52.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
scipy>=1.11.0
